Create flashcards that belong to stack and show
The user now has the ability to create flashcards and assign them to a
specific stack. The user can now get a list of all of the created
flashcards.

BUG: Reloading the page saves the stack information in the dropdown
menu but it does not save the flashcard information correctly, the user
cannot add or show flashcards if the page is reloaded.