Fix saveData and a few other bugs
User can now reliably save all information on the site, this is done
automatically before the site is unloaded. Removed several saving points
in favor of saving when the user exits the program. Still need to work
on finding the best way to present the cards on the home page - do not
want the user to have to click a button to refresh the list.