Create 'Review' functionality
Fleshed out review page which allows the user to sort and page through
flashcards. In this view the user does not have the ability to add or
remove cards/stacks, just the ability to view them one at a time using
buttons which show next/prev flashcards from the selected stack.

Also now reset the home and review pages when switching between them.